Split Database\Table into abstract base class and Comment_Author_Table
