+++ includes/avatar-privacy/cli/class-uninstall-command.php
@@ -29,7 +29,7 @@ namespace Avatar_Privacy\CLI;
 use Avatar_Privacy\CLI\Abstract_Command;
 use Avatar_Privacy\Components\Setup;
 use Avatar_Privacy\Components\Uninstallation;
-use Avatar_Privacy\Data_Storage\Database\Table as Database;
+use Avatar_Privacy\Data_Storage\Database\Comment_Author_Table as Database;
 
 use WP_CLI;
