+++ tests/avatar-privacy/cli/class-database-command-test.php
@@ -35,7 +35,7 @@ use Mockery as m;
 use Avatar_Privacy\CLI\Database_Command;
 
 use Avatar_Privacy\Core;
-use Avatar_Privacy\Data_Storage\Database\Table as Database;
+use Avatar_Privacy\Data_Storage\Database\Comment_Author_Table as Database;
 
 /**
  * Avatar_Privacy\CLI\Database_Command unit test.
