+++ tests/avatar-privacy/cli/class-uninstall-command-test.php
@@ -36,7 +36,7 @@ use Avatar_Privacy\CLI\Uninstall_Command;
 
 use Avatar_Privacy\Components\Setup;
 use Avatar_Privacy\Components\Uninstallation;
-use Avatar_Privacy\Data_Storage\Database\Table as Database;
+use Avatar_Privacy\Data_Storage\Database\Comment_Author_Table as Database;
 
 /**
  * Avatar_Privacy\CLI\Uninstall_Command unit test.
