+++ tests/avatar-privacy/data-storage/database/class-table-test.php
@@ -34,9 +34,7 @@ use Mockery as m;
 
 use org\bovigo\vfs\vfsStream;
 
-use Avatar_Privacy\Core\Hasher;
 use Avatar_Privacy\Data_Storage\Database\Table;
-use Avatar_Privacy\Data_Storage\Network_Options;
 
 /**
  * Avatar_Privacy\Data_Storage\Database\Table unit test.
@@ -55,19 +53,18 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 	 */
 	private $sut;
 
-	/**
-	 * Helper object.
-	 *
-	 * @var Hasher
-	 */
-	private $hasher;
+	const TABLE_BASENAME   = 'my_table';
+	const UPDATE_THRESHOLD = '0.5';
 
-	/**
-	 * Helper object.
-	 *
-	 * @var Network_Options
-	 */
-	private $network_options;
+	// Re-use actual fields from Column_Author_Table to keep test cases close to reality.
+	const COLUMN_FORMATS = [
+		'id'           => '%d',
+		'email'        => '%s',
+		'hash'         => '%s',
+		'use_gravatar' => '%d',
+		'last_updated' => '%s',
+		'log_message'  => '%s',
+	];
 
 	/**
 	 * Sets up the fixture, for example, opens a network connection.
@@ -91,14 +88,11 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 		];
 
 		// Set up virtual filesystem.
-		$root = vfsStream::setup( 'root', null, $filesystem );
+		vfsStream::setup( 'root', null, $filesystem );
 		set_include_path( 'vfs://root/' ); // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.runtime_configuration_set_include_path
 
-		$this->hasher          = m::mock( Hasher::class );
-		$this->network_options = m::mock( Network_Options::class );
-
 		// Partially mock system under test.
-		$this->sut = m::mock( Table::class, [ $this->hasher, $this->network_options ] )->makePartial()->shouldAllowMockingProtectedMethods();
+		$this->sut = m::mock( Table::class, [ self::TABLE_BASENAME, self::UPDATE_THRESHOLD, self::COLUMN_FORMATS ] )->makePartial()->shouldAllowMockingProtectedMethods();
 	}
 
 	/**
@@ -107,12 +101,19 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 	 * @covers ::__construct
 	 */
 	public function test_constructor() {
-		$mock = m::mock( Table::class )->makePartial();
-		$mock->__construct( $this->hasher, $this->network_options );
+		$table_basename   = 'foobar';
+		$update_threshold = '0.6';
+		$column_formats   = [
+			'foobar' => '%s',
+			'foobaz' => '%d',
+		];
 
-		$this->assert_attribute_same( $this->hasher, 'hasher', $mock );
-		$this->assert_attribute_same( $this->network_options, 'network_options', $mock );
-		$this->assert_is_array( $this->get_value( $mock, 'placeholder' ) );
+		$mock = m::mock( Table::class )->makePartial();
+		$mock->__construct( $table_basename, $update_threshold, $column_formats );
+
+		$this->assert_attribute_same( $table_basename, 'table_basename', $mock );
+		$this->assert_attribute_same( $update_threshold, 'update_threshold', $mock );
+		$this->assert_attribute_same( $column_formats, 'column_formats', $mock );
 	}
 
 	/**
@@ -156,10 +157,11 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 	 */
 	public function test_get_table_name() {
 		$site_id = 5;
+		$prefix  = 'my_prefix_';
 
-		$this->sut->shouldReceive( 'get_table_prefix' )->once()->with( $site_id )->andReturn( 'prefix_' );
+		$this->sut->shouldReceive( 'get_table_prefix' )->once()->with( $site_id )->andReturn( $prefix );
 
-		$this->assertSame( 'prefix_avatar_privacy', $this->sut->get_table_name( $site_id ) );
+		$this->assertSame( $prefix . self::TABLE_BASENAME, $this->sut->get_table_name( $site_id ) );
 	}
 
 	/**
@@ -192,35 +194,6 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 		$wpdb->shouldReceive( 'get_var' )->once()->with( 'PREPARED_SQL' )->andReturn( null );
 
 		$this->assertFalse( $this->sut->table_exists( $table_name ) );
-	}
-
-	/**
-	 * Provides data for testing use_global_table.
-	 *
-	 * @return array
-	 */
-	public function provide_use_global_table_data() {
-		return [
-			[ true ],
-			[ false ],
-		];
-	}
-
-	/**
-	 * Tests ::use_global_table.
-	 *
-	 * @covers ::use_global_table
-	 *
-	 * @dataProvider provide_use_global_table_data
-	 *
-	 * @param bool $result The expected result.
-	 */
-	public function test_use_global_table( $result ) {
-		$this->network_options->shouldReceive( 'get' )->once()->with( Network_Options::USE_GLOBAL_TABLE, false )->andReturn( $result );
-
-		Filters\expectApplied( 'avatar_privacy_enable_global_table' )->once()->with( $result );
-
-		$this->assertSame( $result, $this->sut->use_global_table( $result ) );
 	}
 
 	/**
@@ -260,12 +233,14 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 		// Global objects.
 		$wpdb = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 		if ( $property_exists ) {
-			$wpdb->avatar_privacy = $table_name;
+			$table_basename        = self::TABLE_BASENAME;
+			$wpdb->$table_basename = $table_name;
 		}
 
 		if ( ! $update && $property_exists ) {
 			$this->sut->shouldReceive( 'get_table_name' )->never();
 			$this->sut->shouldReceive( 'table_exists' )->never();
+			$this->sut->shouldReceive( 'get_table_definition' )->never();
 			$this->sut->shouldReceive( 'db_delta' )->never();
 			$this->sut->shouldReceive( 'register_table' )->never();
 		} else {
@@ -273,8 +248,10 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 
 			if ( ! $update && $table_exists ) {
 				$this->sut->shouldReceive( 'table_exists' )->once()->with( $table_name )->andReturn( $table_exists );
+				$this->sut->shouldReceive( 'get_table_definition' )->never();
 				$this->sut->shouldReceive( 'db_delta' )->never();
 			} else {
+				$this->sut->shouldReceive( 'get_table_definition' )->once()->with( $table_name )->andReturn( 'SQL string' );
 				$wpdb->shouldReceive( 'get_charset_collate' )->once()->andReturn( 'my_collation' );
 				$this->sut->shouldReceive( 'db_delta' )->once()->with( m::type( 'string' ) );
 
@@ -310,6 +287,7 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 
 		$wpdb->shouldReceive( 'get_charset_collate' )->once()->andReturn( 'my_collation' );
 
+		$this->sut->shouldReceive( 'get_table_definition' )->once()->with( $table_name )->andReturn( 'SQL string' );
 		$this->sut->shouldReceive( 'db_delta' )->once()->with( m::type( 'string' ) );
 
 		$this->sut->shouldReceive( 'register_table' )->never();
@@ -337,9 +315,9 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 
 		$this->assertNull( $this->sut->register_table( $db, $table_name ) );
 
-		$this->assert_attribute_contains( Table::TABLE_BASENAME, 'tables', $db );
-		$this->assert_attribute_not_contains( Table::TABLE_BASENAME, 'ms_global_tables', $db );
-		$this->assert_attribute_same( $table_name, Table::TABLE_BASENAME, $db );
+		$this->assert_attribute_contains( self::TABLE_BASENAME, 'tables', $db );
+		$this->assert_attribute_not_contains( self::TABLE_BASENAME, 'ms_global_tables', $db );
+		$this->assert_attribute_same( $table_name, self::TABLE_BASENAME, $db );
 	}
 
 	/**
@@ -360,9 +338,9 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 
 		$this->assertNull( $this->sut->register_table( $db, $table_name ) );
 
-		$this->assert_attribute_contains( Table::TABLE_BASENAME, 'tables', $db );
-		$this->assert_attribute_not_contains( Table::TABLE_BASENAME, 'ms_global_tables', $db );
-		$this->assert_attribute_same( $table_name, Table::TABLE_BASENAME, $db );
+		$this->assert_attribute_contains( self::TABLE_BASENAME, 'tables', $db );
+		$this->assert_attribute_not_contains( self::TABLE_BASENAME, 'ms_global_tables', $db );
+		$this->assert_attribute_same( $table_name, self::TABLE_BASENAME, $db );
 	}
 
 	/**
@@ -383,9 +361,9 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 
 		$this->assertNull( $this->sut->register_table( $db, $table_name ) );
 
-		$this->assert_attribute_not_contains( Table::TABLE_BASENAME, 'tables', $db );
-		$this->assert_attribute_contains( Table::TABLE_BASENAME, 'ms_global_tables', $db );
-		$this->assert_attribute_same( $table_name, Table::TABLE_BASENAME, $db );
+		$this->assert_attribute_not_contains( self::TABLE_BASENAME, 'tables', $db );
+		$this->assert_attribute_contains( self::TABLE_BASENAME, 'ms_global_tables', $db );
+		$this->assert_attribute_same( $table_name, self::TABLE_BASENAME, $db );
 	}
 
 	/**
@@ -429,661 +407,37 @@ class Table_Test extends \Avatar_Privacy\Tests\TestCase {
 	}
 
 	/**
-	 * Tests ::migrate_from_global_table.
+	 * Tests ::get_format.
 	 *
-	 * @covers ::migrate_from_global_table
+	 * @covers ::get_format
 	 */
-	public function test_migrate_from_global_table() {
-		global $wpdb;
-		$wpdb              = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$site_id           = 7;
-		$main_site_id      = 1;
-		$network_id        = 5;
-		$table_name        = 'my_custom_table';
-		$global_table_name = 'my_global_table';
-		$result            = 4;
-
-		// Interim results.
-		$global_id_1 = 2;
-		$global_id_2 = 4;
-		$global_id_3 = 5;
-		$global_id_4 = 13;
-		$email_1     = 'foo@bar.org';
-		$email_2     = 'foobar@bar.org';
-		$email_3     = 'bar@foo.org';
-		$email_4     = 'x@foobar.org';
-		$local_id_1  = 3;
-		$local_id_2  = 11;
-
-		$rows_to_migrate = [
-			$global_id_1 => (object) [
-				'id'           => $global_id_1,
-				'email'        => $email_1,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			$global_id_2 => (object) [
-				'id'           => $global_id_2,
-				'email'        => $email_2,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-			$global_id_3 => (object) [
-				'id'           => $global_id_3,
-				'email'        => $email_3,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			$global_id_4 => (object) [
-				'id'           => $global_id_4,
-				'email'        => $email_4,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
+	public function test_get_format() {
+		$columns  = [
+			'log_message'  => 'foo',
+			'use_gravatar' => 1,
+			'hash'         => 'bar',
 		];
-		$emails          = [
-			$global_id_1 => $email_1,
-			$global_id_2 => $email_2,
-			$global_id_3 => $email_3,
-			$global_id_4 => $email_4,
-		];
-		$existing_rows   = [
-			$local_id_1  => (object) [
-				'id'           => $local_id_1,
-				'email'        => $email_1,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-17 22:23:08',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			$local_id_2  => (object) [
-				'id'           => $local_id_2,
-				'email'        => $email_4,
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-19 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-		];
+		$expected = [ '%s', '%d', '%s' ];
 
-		Functions\expect( 'get_main_site_id' )->once()->andReturn( $main_site_id );
-
-		$this->sut->shouldReceive( 'get_table_name' )->once()->with( $site_id )->andReturn( $table_name );
-		$this->sut->shouldReceive( 'get_table_name' )->once()->with( $main_site_id )->andReturn( $global_table_name );
-
-		$wpdb->shouldReceive( 'esc_like' )->once()->with( $site_id )->andReturn( $site_id );
-		$wpdb->shouldReceive( 'prepare' )->once()->with( "SELECT * FROM `{$global_table_name}` WHERE log_message LIKE %s", "set with comment % (site: %, blog: {$site_id})" )->andReturn( 'SELECT_QUERY' );
-		$wpdb->shouldReceive( 'get_results' )->once()->with( 'SELECT_QUERY', \OBJECT_K )->andReturn( $rows_to_migrate );
-
-		Functions\expect( 'wp_list_pluck' )->once()->with( $rows_to_migrate, 'email', 'id' )->andReturn( $emails );
-		$this->sut->shouldReceive( 'prepare_email_query' )->once()->with( $emails, $table_name )->andReturn( 'EMAIL_QUERY' );
-		$wpdb->shouldReceive( 'get_results' )->once()->with( 'EMAIL_QUERY' )->andReturn( $existing_rows );
-
-		$this->sut->shouldReceive( 'prepare_insert_update_query' )->once()->with(
-			m::on(
-				function( $a ) use ( $local_id_1 ) {
-					return \array_keys( $a ) === [ $local_id_1 ];
-				}
-			),
-			m::on(
-				function( $a ) use ( $global_id_2, $global_id_3 ) {
-					return \array_keys( $a ) === [ $global_id_2, $global_id_3 ];
-				}
-			),
-			$table_name
-		)->andReturn( 'INSERT_UPDATE_QUERY' );
-		$wpdb->shouldReceive( 'query' )->once()->with( 'INSERT_UPDATE_QUERY' )->andReturn( 4 );
-
-		$this->sut->shouldReceive( 'prepare_delete_query' )->once()->with(
-			m::on(
-				function( $a ) use ( $global_id_1, $global_id_2, $global_id_3, $global_id_4 ) {
-					// ID1 and ID4 exist, therefore they come first here.
-					// ID2 and ID3 are migrated.
-					return [ $global_id_1, $global_id_4, $global_id_2, $global_id_3 ] === $a;
-				}
-			),
-			$global_table_name
-		)->andReturn( 'DELETE_QUERY' );
-		$wpdb->shouldReceive( 'query' )->once()->with( 'DELETE_QUERY' )->andReturn( 4 );
-
-		$this->assertSame( $result, $this->sut->migrate_from_global_table( $site_id ) );
+		$this->assertSame( $expected, $this->sut->get_format( $columns ) );
 	}
 
 	/**
-	 * Tests ::migrate_from_global_table.
+	 * Tests ::get_format.
 	 *
-	 * @covers ::migrate_from_global_table
+	 * @covers ::get_format
 	 */
-	public function test_migrate_from_global_table_same_table_name() {
-		global $wpdb;
-		$wpdb              = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$site_id           = 2;
-		$main_site_id      = 2;
-		$network_id        = 5;
-		$table_name        = 'my_global_table';
-		$global_table_name = 'my_global_table';
-		$result            = false;
-
-		Functions\expect( 'get_main_site_id' )->once()->andReturn( $main_site_id );
-
-		$this->sut->shouldReceive( 'get_table_name' )->once()->with( $site_id )->andReturn( $table_name );
-		$this->sut->shouldReceive( 'get_table_name' )->once()->with( $main_site_id )->andReturn( $global_table_name );
-
-		$this->assertSame( $result, $this->sut->migrate_from_global_table( $site_id ) );
-	}
-
-	/**
-	 * Tests ::prepare_insert_update_query.
-	 *
-	 * @covers ::prepare_insert_update_query
-	 *
-	 * @uses Avatar_Privacy\Data_Storage\Table::get_placeholders
-	 * @uses Avatar_Privacy\Data_Storage\Table::get_prepared_values
-	 * @uses Avatar_Privacy\Data_Storage\Table::get_update_clause
-	 */
-	public function test_prepare_insert_update_query() {
-		global $wpdb;
-		$wpdb            = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$network_id      = 5;
-		$site_id         = 3;
-		$rows_to_update  = [
-			3  => (object) [
-				'id'           => 1,
-				'email'        => 'foo@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-17 22:23:08',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			11 => (object) [
-				'id'           => 7,
-				'email'        => 'xxx@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-19 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
+	public function test_get_format_invalid_column() {
+		$columns  = [
+			'log_message'  => 'foo',
+			'use_gravatar' => 1,
+			'hash'         => 'bar',
+			'foo'          => 'bar',
 		];
-		$update_count    = \count( $rows_to_update );
-		$rows_to_migrate = [
-			(object) [
-				'id'           => 32,
-				'email'        => 'foobar@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-			(object) [
-				'id'           => 33,
-				'email'        => 'bar@foo.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			(object) [
-				'id'           => 66,
-				'email'        => 'x@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-		];
-		$migrate_count   = \count( $rows_to_migrate );
-		$total_count     = $update_count + $migrate_count;
-		$table_name      = 'my_table';
-		$result          = 'INSERT_UPDATE_QUERY';
+		$expected = [ '%s', '%d', '%s' ];
 
-		$wpdb->shouldReceive( 'prepare' )->once()->with( m::pattern( "/INSERT INTO `{$table_name}` \(id,email,hash,use_gravatar,last_updated,log_message\)\s+VALUES (\(%d,%s,%s,%d,%s,%s\),?){{$update_count}}(\(NULL,%s,%s,%d,%s,%s\),?){{$migrate_count}}\s+ON DUPLICATE KEY UPDATE\s+email = VALUES\(email\),\s+hash = VALUES\(hash\),\s+use_gravatar = VALUES\(use_gravatar\),\s+last_updated = VALUES\(last_updated\),\s+log_message = VALUES\(log_message\)/mu" ), m::type( 'array' ) )->andReturn( $result );
+		$this->expectException( \RuntimeException::class );
 
-		$this->assertSame( $result, $this->sut->prepare_insert_update_query( $rows_to_update, $rows_to_migrate, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_insert_update_query.
-	 *
-	 * @covers ::prepare_insert_update_query
-	 */
-	public function test_prepare_insert_update_query_no_rows() {
-		global $wpdb;
-		$wpdb            = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$rows_to_update  = [];
-		$rows_to_migrate = [];
-		$table_name      = 'my_table';
-		$result          = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_insert_update_query( $rows_to_update, $rows_to_migrate, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_insert_update_query.
-	 *
-	 * @covers ::prepare_insert_update_query
-	 */
-	public function test_prepare_insert_update_query_empty_table_name() {
-		global $wpdb;
-		$wpdb            = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$network_id      = 5;
-		$site_id         = 3;
-		$rows_to_update  = [
-			3  => (object) [
-				'id'           => 1,
-				'email'        => 'foo@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-17 22:23:08',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			11 => (object) [
-				'id'           => 7,
-				'email'        => 'xxx@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-19 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-		];
-		$rows_to_migrate = [
-			(object) [
-				'id'           => 32,
-				'email'        => 'foobar@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-			(object) [
-				'id'           => 33,
-				'email'        => 'bar@foo.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			(object) [
-				'id'           => 66,
-				'email'        => 'x@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-		];
-		$table_name      = '';
-		$result          = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_insert_update_query( $rows_to_update, $rows_to_migrate, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_insert_update_query.
-	 *
-	 * @covers ::prepare_insert_update_query
-	 */
-	public function test_prepare_insert_update_query_no_valid_fields() {
-		global $wpdb;
-		$wpdb            = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$network_id      = 5;
-		$site_id         = 3;
-		$rows_to_update  = [
-			3  => (object) [
-				'id'           => 1,
-				'email'        => 'foo@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-17 22:23:08',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			11 => (object) [
-				'id'           => 7,
-				'email'        => 'xxx@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-19 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-		];
-		$rows_to_migrate = [
-			(object) [
-				'id'           => 32,
-				'email'        => 'foobar@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-			(object) [
-				'id'           => 33,
-				'email'        => 'bar@foo.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			(object) [
-				'id'           => 66,
-				'email'        => 'x@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-18 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 0,
-			],
-		];
-		$table_name      = 'foobar';
-		$fields          = [ 'foobar' ];
-		$result          = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_insert_update_query( $rows_to_update, $rows_to_migrate, $table_name, $fields ) );
-	}
-
-	/**
-	 * Provides data for testing get_update_clause.
-	 *
-	 * @return array
-	 */
-	public function provide_get_update_clause_data() {
-		return [
-			[
-				[ 'email', 'hash', 'use_gravatar' ],
-				"email = VALUES(email),\nhash = VALUES(hash),\nuse_gravatar = VALUES(use_gravatar),\nlast_updated = last_updated,\nlog_message = log_message",
-			],
-			[
-				[ 'log_message' ],
-				"email = email,\nhash = hash,\nuse_gravatar = use_gravatar,\nlast_updated = last_updated,\nlog_message = VALUES(log_message)",
-			],
-			[
-				[ 'email', 'hash', 'use_gravatar', 'last_updated', 'log_message' ],
-				"email = VALUES(email),\nhash = VALUES(hash),\nuse_gravatar = VALUES(use_gravatar),\nlast_updated = VALUES(last_updated),\nlog_message = VALUES(log_message)",
-			],
-
-		];
-	}
-
-	/**
-	 * Tests ::get_update_clause.
-	 *
-	 * @covers ::get_update_clause
-	 *
-	 * @dataProvider provide_get_update_clause_data
-	 *
-	 * @param string[] $fields  A list of columns.
-	 * @param string   $result  The expected result.
-	 */
-	public function test_get_update_clause( array $fields, $result ) {
-		$this->assertSame( $result, $this->sut->get_update_clause( $fields ) );
-	}
-
-	/**
-	 * Tests ::get_prepared_values.
-	 *
-	 * @covers ::get_prepared_values
-	 */
-	public function test_get_prepared_values() {
-		$network_id = 5;
-		$site_id    = 3;
-		$rows       = [
-			3  => (object) [
-				'id'           => 1,
-				'email'        => 'foo@bar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-17 22:23:08',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-			11 => (object) [
-				'id'           => 7,
-				'email'        => 'xxx@foobar.org',
-				'hash'         => 'hash',
-				'last_updated' => '2018-12-19 10:00:00',
-				'log_message'  => "set with comment 8 (site: {$network_id}, blog: {$site_id})",
-				'use_gravatar' => 1,
-			],
-		];
-		$fields     = [
-			'email',
-			'use_gravatar',
-			'last_updated',
-		];
-		$result     = [
-			3, // real ID.
-			'foo@bar.org',
-			1, // use_gravatar.
-			'2018-12-17 22:23:08',
-			11, // real ID.
-			'xxx@foobar.org',
-			1, // use_gravatar.
-			'2018-12-19 10:00:00',
-		];
-
-		$this->assertSame( $result, $this->sut->get_prepared_values( $rows, $fields, true ) );
-	}
-
-	/**
-	 * Provides data for testing get_placeholders.
-	 *
-	 * @return array
-	 */
-	public function provide_get_placeholders_data() {
-		return [
-			[ [ 'email', 'hash', 'use_gravatar' ], true, '(%d,%s,%s,%d)' ],
-			[ [ 'email', 'hash', 'use_gravatar' ], false, '(NULL,%s,%s,%d)' ],
-			[ [ 'email', 'hash', 'use_gravatar', 'last_updated', 'log_message' ], true, '(%d,%s,%s,%d,%s,%s)' ],
-		];
-	}
-
-	/**
-	 * Tests ::get_placeholders.
-	 *
-	 * @covers ::get_placeholders
-	 *
-	 * @dataProvider provide_get_placeholders_data
-	 *
-	 * @param string[] $fields  A list of columns.
-	 * @param bool     $with_id Whether the ID should be included.
-	 * @param string   $result  The expected result.
-	 */
-	public function test_get_placeholders( array $fields, $with_id, $result ) {
-		$this->assertSame( $result, $this->sut->get_placeholders( $fields, $with_id ) );
-	}
-
-	/**
-	 * Tests ::prepare_email_query.
-	 *
-	 * @covers ::prepare_email_query
-	 */
-	public function test_prepare_email_query() {
-		global $wpdb;
-		$wpdb        = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$emails      = [ 'foo@bar.org', 'foobar@example.org', 'bar@foo.com' ];
-		$email_count = \count( $emails );
-		$table_name  = 'my_table';
-		$result      = 'EMAILS_QUERY';
-
-		$wpdb->shouldReceive( 'prepare' )->once()->with( m::pattern( "/SELECT \* FROM `{$table_name}` WHERE email IN \((%s,?){{$email_count}}\)/" ), $emails )->andReturn( $result );
-
-		$this->assertSame( $result, $this->sut->prepare_email_query( $emails, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_email_query.
-	 *
-	 * @covers ::prepare_email_query
-	 */
-	public function test_prepare_email_query_no_emails() {
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$emails     = [];
-		$table_name = 'my_table';
-		$result     = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_email_query( $emails, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_email_query.
-	 *
-	 * @covers ::prepare_email_query
-	 */
-	public function test_prepare_email_query_empty_table_name() {
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$emails     = [ 'foo@bar.org', 'foobar@example.org', 'bar@foo.com' ];
-		$table_name = '';
-		$result     = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_email_query( $emails, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_delete_query.
-	 *
-	 * @covers ::prepare_delete_query
-	 */
-	public function test_prepare_delete_query() {
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$ids        = [ 5, 10, 11, 17 ];
-		$id_count   = \count( $ids );
-		$table_name = 'my_table';
-		$result     = 'DELETE_QUERY';
-
-		$wpdb->shouldReceive( 'prepare' )->once()->with( m::pattern( "/DELETE FROM `{$table_name}` WHERE id IN \((%d,?){{$id_count}}\)/" ), $ids )->andReturn( $result );
-
-		$this->assertSame( $result, $this->sut->prepare_delete_query( $ids, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_delete_query.
-	 *
-	 * @covers ::prepare_delete_query
-	 */
-	public function test_prepare_delete_query_no_ids() {
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$ids        = [];
-		$table_name = 'my_table';
-		$result     = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_delete_query( $ids, $table_name ) );
-	}
-
-	/**
-	 * Tests ::prepare_delete_query.
-	 *
-	 * @covers ::prepare_delete_query
-	 */
-	public function test_prepare_delete_query_empty_table_name() {
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$ids        = [ 5, 10, 11, 17 ];
-		$table_name = '';
-		$result     = false;
-
-		$wpdb->shouldReceive( 'prepare' )->never();
-
-		$this->assertSame( $result, $this->sut->prepare_delete_query( $ids, $table_name ) );
-	}
-
-	/**
-	 * Tests ::maybe_upgrade_table_data.
-	 *
-	 * @covers ::maybe_upgrade_table_data
-	 */
-	public function test_maybe_upgrade_table_data() {
-		$network_id = 5;
-		$site_id    = 3;
-		$rows       = [
-			3  => (object) [
-				'id'           => 3,
-				'email'        => 'foo@bar.org',
-			],
-			11 => (object) [
-				'id'           => 11,
-				'email'        => 'xxx@foobar.org',
-			],
-		];
-		$fields     = [
-			'email',
-			'use_gravatar',
-			'last_updated',
-		];
-		$result     = 2; // Affected row count.
-
-		// Fake global.
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$table_name = 'my_table';
-
-		$this->sut->shouldReceive( 'get_table_name' )->once()->andReturn( $table_name );
-
-		$wpdb->shouldReceive( 'get_results' )->once()->with( "SELECT id, email FROM {$table_name} WHERE hash is null", \OBJECT_K )->andReturn( $rows );
-		$this->hasher->shouldReceive( 'get_hash' )->times( \count( $rows ) )->with( m::type( 'string' ) )->andReturn( 'hashed email' );
-
-		$this->sut->shouldReceive( 'prepare_insert_update_query' )->once()->with( $rows, [], $table_name, [ 'hash' ] )->andReturn( 'UPDATE_QUERY' );
-		$wpdb->shouldReceive( 'query' )->once()->with( 'UPDATE_QUERY' );
-
-		$this->assertSame( $result, $this->sut->maybe_upgrade_table_data() );
-	}
-
-	/**
-	 * Tests ::maybe_upgrade_table_data.
-	 *
-	 * @covers ::maybe_upgrade_table_data
-	 */
-	public function test_maybe_upgrade_table_data_error() {
-		$network_id = 5;
-		$site_id    = 3;
-		$rows       = [
-			3  => (object) [
-				'id'           => 3,
-				'email'        => 'foo@bar.org',
-			],
-			11 => (object) [
-				'id'           => 11,
-				'email'        => 'xxx@foobar.org',
-			],
-		];
-		$fields     = [
-			'email',
-			'use_gravatar',
-			'last_updated',
-		];
-
-		// Fake global.
-		global $wpdb;
-		$wpdb       = m::mock( 'wpdb' ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
-		$table_name = 'my_table';
-
-		$this->sut->shouldReceive( 'get_table_name' )->once()->andReturn( $table_name );
-
-		$wpdb->shouldReceive( 'get_results' )->once()->with( "SELECT id, email FROM {$table_name} WHERE hash is null", \OBJECT_K )->andReturn( $rows );
-		$this->hasher->shouldReceive( 'get_hash' )->times( \count( $rows ) )->with( m::type( 'string' ) )->andReturn( 'hashed email' );
-
-		$this->sut->shouldReceive( 'prepare_insert_update_query' )->once()->with( $rows, [], $table_name, [ 'hash' ] )->andReturn( false );
-		$wpdb->shouldReceive( 'query' )->never();
-
-		$this->assertSame( 0, $this->sut->maybe_upgrade_table_data() );
+		$this->assertSame( $expected, $this->sut->get_format( $columns ) );
 	}
 }
